+++ app/src/main/java/com/github/brandon_patterson/chemist/matter/NoSuchElementException.java
@@ -1,0 +1,9 @@
+package com.github.brandon_patterson.chemist.matter;
+
+public class NoSuchElementException extends Exception
+{
+    public NoSuchElementException(String message)
+    {
+        super(message);
+    }
+}