+++ CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - license (MIT)
 - basic CI (for unit tests only)
 - homepage
+- Elements enum (symbol/name/number only)